requirements.txt
python-telegram-bot==13.1
requests==2.25.0
beautifulsoup4==4.9.3
soupsieve==2.0.1
pymongo==3.11.2
dnspython==2.0.0